-change id reference to integer

+++ src/main/java/com/ipayafrica/elipapower/service/IErrorCodeService.java
@@ -2,7 +2,7 @@ package com.ipayafrica.elipapower.service;
 
 import com.ipayafrica.elipapower.model.ErrorCode;
 
-public interface IErrorCodeService extends IGenericService<ErrorCode, Long> {
-	ErrorCode getByErrorCodeID (Long errorCodeId);
+public interface IErrorCodeService extends IGenericService<ErrorCode, Integer> {
+	ErrorCode getByErrorCodeID (Integer errorCodeId);
 	int getByMessageCode (String messageCode);
 }
